pyproject.toml
[tool.poetry]
name = "todolist-package"
version = "0.1.0"
description = ""
authors = ["Dae-Jeong <v4chelsea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
